refactor: reuse read_serial fn with nusb
